Enabled Fortify for libruntime.a
There is currently no code in libruntime.a that Fortify can protect so this
doesn't actually change the generated assembly. However, we may add code in the
future that Fortify can protect.

clang defines `_FORTIFY_SOURCE` by default so we need to undefine then redefine
it to avoid build errors.