WIP tree-sitter grammar development tips doc
Started a very WIP doc that aims to bring a bit more empiricism to the process of grammar development. This doc is supposed to serve as a `parse tree style guide` as well as a `grammar development guide`. By pairing with Max (and others), I've learned how to exploit the mechanics of tree-sitter without really knowing a lot of underlying logic or when to make certain decisions. Since it's not immediately intuitive for me, I wanted to document my learnings and rely on you all to polish this.

@maxbrunsfeld, @tclem, @rewinfrey, @robrix, @joshvera — would love your expertise here. It would be nice to put some writing down and not simply rely on word of mouth as the means to transition this knowledge, especially as we hire. If someone wants to pair with me later on this, I'm happy to do that too.  Additionally, if folks have a good understanding of what the tree structure should be like for the purpose of abstract interpretation (looking at you @tclem) please chime in!