Add API for indicating which external tokens can match the empty string
Previously, when evaluating whether parse state merges were legal,
the ParseTableBuilder would assume that any external token could
potentially match the empty string. This logic was added to fix some
invalid parse state merging that was happening when generating
the bash parser.

Unfortunately, it resulted in overly conservative parse state merging,
which was especially undesirable in a language like Ruby with a large
number of parse states and a lot of external tokens which *cannot*
actually match the empty string.

Now, external tokens which match the empty string should explicitly
be marked as such using syntax like this:

```
externals: $ => [
  $.foo,
  $.bar,
  [$.concat, {blank: true},
]
```

I wish it weren't necessary to expose this detail in the grammar API,
but reducing the size of the ruby parser is very important and I don't
see any way to achieve this without requiring the grammar author to
explicitly indicate which external tokens have this property.