Add criterion benchmarks
The previous benchmark system was unreliable since it did not attempt to iterate on the unit under test. The criterion benchmark library, on the other hand, produces useful statistical results for the benchmarks and allows for reliable, statistically significant comparisons. However, the criterion benchmarks take significantly longer to run (~9 hours vs. ~9 seconds) and therefore are not used by default.

 * .gitignore: ignore CLI target output files produced by critereon
 * Cargo.lock: update lock file
 * cli/Cargo.toml: add criterion to benchmarks and dev dependencies
 * cli/benches/crit-benchmark.rs: this is based on the original benchmarks.rs but uses criterion and makes some functions for reading the input
 * script/benchmark: specify which benchmark to run and default to the original benchmark
 * script/benchmark.cmd: likewise