Adding API to get field name of a named TSNode
This PR adds an API to get name of the field of TSNode's named child.
It uses same set of arguments as that of ts_node_field_name_for_child, but returns
field name if it is found, otherwise it returns NULL.

This API is useful to implement custom printing of S-expressions such
as following:

"(binary_expression
(binary_expression_left (identifier))
(binary_expression_operator ("+"))
(binary_expression_right (identifier)
)"

The added API's purpose is same as that of ts_node_field_name_for_child,
but differs in the handling of named child. ts_node_field_name_for_child
does not support named child.