Add WGSL WebGPU Shading Language
This PR adds WebGPU Shading Language to the list of available grammars, the linked repository's grammar is periodically automatically extracted from the WGSL specification itself, which is actually extracted every single time where the specification gets a modification in itself to check the validity of both the syntax and the examples. Thank you very much for developing and maintaining tree-sitter, fantastic project!