cicd: switch main workflow to work on any branch except `release/v*` prefixed
The `release/v*` prefixed branches can be used only with a PR opened on
such branches. The release branches only accepted in the main repo
and not from forks.

Also this workflow enables runining on every `push` unconditionaly.
This leads to parallel runs on `push` and `pull_request` events
and that would seen in PR checks also in parallel, but this is
useful because checks that run on the `push` event runs on
head ref while checks on the `pull_request` event runs on
an additional `refs/pull/<PR>/merge` ref what is created
by Github on every push to a PR to check that it can be
merged to a target branch without conflicts and passed checks.