feat(lexer)!: `TSLexer.get_byte()`
Currently, since external scanners are called on-demand at arbitrary
points, they are limited in context/state.

An example of where this leads to redundant computation is when
determining the effective indent of Python comments, which depends not
only on the comment's indentation but also the indentation of the next
non-comment line, which is an expensive lookahead (O(comment_size)). If
there are many consecutive comments (e.g. as generated by an IDE when
commenting out a block of code), then this computation is redundant, and
in total quadratic.

`TSLexer.get_byte()` improves the statefulness of external scanners,
allowing to keep track of lookahead state, reducing redundancy.

BREAKING CHANGE: This breaks the `TSLexer` ABI. To minimize disruption,
it can be bundled with other breaking changes.