ci: introduce "response bot"
It's basically a bot that closes issues or PRs if the author hasn't
responded after 30 days. This is only done on issues/PRs that has the
`more-information-needed` label and is meant to be used for issues that
has hit a dead-end and cannot progress without further information from
the author. If the author responds after an issue is closed and gives
more context the issue can simply be reopened.