Cfg unix or wasi for fd in Rust Bindings
not(windows) includes non-wasi web assembly which causes issues when using the rust bindings from non-wasi web assembly.

In lib.rs, `AsRawFd` does not exist for wasm so it was causing the build to fail.

I also changed ffi.rs `_ts_dup` for consistency even though that does not cause a build issue.

This does lead to the situation where if you are not on either of those platforms, print_dot_graphs has no parameter to take a file. While odd, I think that does successfully convey that the current platform is not supported to use that function.