Include "./lib/src" as a header search path in SwiftPM build manifest
After ee9a3c0ebb218990cf391ed987be7f2448c54a73, a part of ICU sources
are vendored in this repository, and the vendored sources include
sibling headers in `#include <unicode/...>` form, so we need to include
the "./lib/src" directory as a header search path.

Apple platform SDKs include ICU headers as a part of system headers, so
it coincidentally works on Apple platforms, but it doesn't work on other
platforms including Linux and WebAssembly.