build: simplify some cmake logic
It's not needed to add logic around the find modules as they generally
behave like you want them, such as skipping a search if the cache
variable is already defined.

Also add only `src/lib.c` to the library since it by design includes the
necessary files to build the library. We could remove `src/lib.c` and
just add the necessary files as it is usually done, but if we already
have such a file it makes sense to use it to ensure it works as
expected.