fix(cli): check more lines in grammar.json loader
Problem: The assumption that `grammar.json` contains the language name
in the first three lines is wrong, since a sizable number of parsers
store an inherited language name character by character. This breaks
building these parsers with `tree-sitter build`.

Solution: Raise the number of lines to 20, which is still O(1) with
respect to grammar size. This can be dropped again once all(!) grammars
are regenerated so the inherited language is stored _after_ the language
name.