allow `data-tree-sitter-highlight` attributes, rather than inline styles
fixes #553

when the `--data-highlight` flag is passed along with the `--html`
flag, the highlighter will inject `data-tree-sitter-highlight`
attributes into the `span` tags, rather than pulling inline styles
from the theme.

this makes it significantly easier to use from downstream tools,
because we can apply css to the output directly, rather than having
the user provide TS with a theme.

only matches which have theme entries are considered for