feat(cli): Support building Rust package for WASM
Tree-sitter claims to support building for WebAssembly and yet on every
step of the way there are road blocks if you want something more than
download a pre-built WASM binary blob. That people struggle with it is
obvious from the various discussions surrounding the matter, e.g.:
- https://github.com/tree-sitter/tree-sitter/discussions/1550
- https://github.com/tree-sitter/tree-sitter/discussions/2010
- https://github.com/tree-sitter/tree-sitter/discussions/1024

It's not just the core library: even languages fail to build for the
target (certainly their Rust bindings). As it turns out it's actually
not that hard to get going and this change boils down the various
recommendations from some of the discussions into making generated
language grammars build for the wasm32-unknown-unknown target out of the
box. All that has to be done is to add and make known a minimal sysroot
with header files that make certain types known and stub out core C
allocation functions.
Note that the sized type definitions are 32 bit specific and will likely
need adjustment once 64 bit WebAssembly is being used more broadly.
Given the wasm32-unknown-unknown check this should be obvious and can be
postponed until wasm64-unknown-unknown [0] is more widely used, I think.

[0] https://doc.rust-lang.org/nightly/rustc/platform-support/wasm64-unknown-unknown.html

Signed-off-by: Daniel Müller <deso@posteo.net>