cli: Adapt default Makefile/CMakeLists for repos not storing grammar.json
This is a new version of #4580, to make it easier not to include
tree-sitter generated files in grammar repos, as outlined in #930.

This makes it possible to run `make` or `cmake --build .` in grammar repos where
`grammar.json` hasn't been included, and still get the parser
compilation to work. The build scripts still work if `grammar.json` is
included.