cli: Adapt default Makefile/CMakeLists for repos not storing grammar.json
This is a step towards making it easier for folks not to include
tree-sitter generated files in grammar repos, as outlined in #930.

This makes it possible to run `make` or `cmake` in grammar repos where
`grammar.json` hasn't been included, and still get the parser
compilation to work. The build scripts still work if `grammar.json` is
included.

One problem that if `grammar.json` isn't include, the generation process
will run twice:
* once as `tree-sitter generate grammar.js`
* a second time as `tree-sitter generate src/grammar.json`
But that's not actually necessary as the first invokation already
generates both `src/grammar.json` and `src/parser.c` (plus other files).

If that's an issue, one could introduce a parameter to `tree-sitter
generate` so that it only does the evaluation from `grammar.js` to
`src/grammar.json` and stops there. This would avoid work getting done
twice.