fix(templates): Generate from grammar.js.
Switch both the template Makefile and CMakeLists.txt to default to
generating from grammar.js instead of grammar.json.

The template does not include a grammar.json, so trying to use the
Makefile or CMake will fail. The docs do suggest running
tree-sitter generate after tree-sitter init, but this shouldn't be
necessary for the build system to work.

Since the documentation focuses on editing the grammar.js file, the
build system should rebuild when changes are made to grammar.js instead
of src/grammar.json